# backend-api/doc/openapiSpec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Cinema Management System API",
    "version": "1.0.0",
    "description": "API for managing cinema operations including movies, cinema rooms, and showtimes",
    "contact": {
      "name": "Cinema Management Team",
      "email": "datb2203438@student.ctu.edu.vn, nhutb2205986@student.ctu.edu.vn"
    }
  },
  "servers": [{
    "url": "http://localhost:3000",
    "description": "Development server"
  }],
...
  "paths": {
...
    "/api/foods": {
      "get": {
        "summary": "Get all foods",
        "description": "Get all food items with pagination and filtering",
        "operationId": "getAllFoods",
        "tags": ["food"],
        "parameters": [{
            "name": "name",
            "in": "query",
            "description": "Filter by food name (partial match)",
            "schema": {
              "type": "string",
              "maxLength": 255
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by food category (partial match)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "is_available",
            "in": "query",
            "description": "Filter by availability status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "min_price",
            "in": "query",
            "description": "Filter by minimum price",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "description": "Filter by maximum price",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of food items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "foods": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Food"
                          }
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/PaginationMetadata"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new food item",
        "description": "Create a new food item (Admin only)",
        "operationId": "createFood",
        "tags": ["food"],
        "security": [{
          "bearerAuth": []
        }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodRequest"
              },
              "example": {
                "name": "Popcorn lớn",
                "description": "Hộp bắp rang lớn, phù hợp cho 2-3 người",
                "price": 65000,
                "image_url": "/public/images/foods/popcorn-large.jpg",
                "category": "Popcorn",
                "is_available": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Food item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "message": {
                      "type": "string",
                      "example": "Food created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "food": {
                          "$ref": "#/components/schemas/Food"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete all food items",
        "description": "Delete all food items (Admin only)",
        "operationId": "deleteAllFoods",
        "tags": ["food"],
        "security": [{
          "bearerAuth": []
        }],
        "responses": {
          "200": {
            "description": "All food items deleted successfully",
            "$ref": "#/components/responses/SuccessNoData"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/foods/{id}": {
      "get": {
        "summary": "Get food item by ID",
        "description": "Get a specific food item by its ID",
        "operationId": "getFoodById",
        "tags": ["food"],
        "parameters": [{
          "$ref": "#/components/parameters/idParam"
        }],
        "responses": {
          "200": {
            "description": "Food item details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "food": {
                          "$ref": "#/components/schemas/Food"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update food item by ID",
        "description": "Update a food item by its ID (Admin only)",
        "operationId": "updateFood",
        "tags": ["food"],
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [{
          "$ref": "#/components/parameters/idParam"
        }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFoodRequest"
              },
              "example": {
                "name": "Popcorn lớn (cập nhật)",
                "price": 70000,
                "is_available": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Food item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "message": {
                      "type": "string",
                      "example": "Food updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "food": {
                          "$ref": "#/components/schemas/Food"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete food item by ID",
        "description": "Delete a food item by its ID (Admin only)",
        "operationId": "deleteFood",
        "tags": ["food"],
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [{
          "$ref": "#/components/parameters/idParam"
        }],
        "responses": {
          "200": {
            "description": "Food item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "message": {
                      "type": "string",
                      "example": "Food deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Get token from /api/auth/login or Google OAuth. Required roles: Admin (full access), Staff (movies/showtimes/customers), Customer (profile only)"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["fail", "error"]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "totalRecords": {
            "type": "integer",
            "minimum": 0,
            "description": "Total number of records"
          },
          "firstPage": {
            "type": "integer",
            "minimum": 1,
            "description": "First page number"
          },
          "lastPage": {
            "type": "integer",
            "minimum": 1,
            "description": "Last page number"
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of records per page"
          }
        }
      },
...
      "Food": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "description": "Food item ID"
          },
          "name": {
            "type": "string",
            "description": "Food item name"
          },
          "description": {
            "type": "string",
            "description": "Food item description"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000000,
            "description": "Food item price in VND"
          },
          "image_url": {
            "type": "string",
            "description": "Food item image URL"
          },
          "category": {
            "type": "string",
            "description": "Food item category"
          },
          "is_available": {
            "type": "boolean",
            "description": "Food item availability status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "CreateFoodRequest": {
        "type": "object",
        "required": ["name", "description", "price", "category"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Food item name"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "description": "Food item description"
          },
          "price": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000000,
            "description": "Food item price in VND"
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "Food item image URL"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Food item category (e.g., Popcorn, Nước uống, Snack, Combo)"
          },
          "is_available": {
            "type": "boolean",
            "default": true,
            "description": "Food item availability status"
          }
        }
      },
      "UpdateFoodRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Food item name"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "description": "Food item description"
          },
          "price": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000000,
            "description": "Food item price in VND"
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "Food item image URL"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Food item category"
          },
          "is_available": {
            "type": "boolean",
            "description": "Food item availability status"
          }
        }
      }
    },
    "responses": {
      "SuccessNoData": {
        "description": "Success response without data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": ["success"]
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "null"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation_error": {
                "value": {
                  "status": "fail",
                  "message": "input.movie_id: Movie ID phải là số dương"
                }
              },
              "invalid_format": {
                "value": {
                  "status": "fail",
                  "message": "Invalid ID format"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Access token is required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "movie_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Movie not found"
                }
              },
              "room_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Room not found"
                }
              },
              "showtime_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Showtime not found"
                }
              },
              "employee_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Employee not found"
                }
              },
              "customer_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Customer not found"
                }
              },
              "booking_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Booking not found"
                }
              },
              "food_not_found": {
                "value": {
                  "status": "fail",
                  "message": "Food not found"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Internal Server Error"
            }
          }
        }
      }
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Resource ID",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "pageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of records per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      }
    }
  }
}